+++ src/imio/helpers/cache.py
@@ -239,3 +239,24 @@ def get_plone_groups_for_user(user_id=None, user=None, the_objects=False):
     else:
         user_groups = user.getGroups()
     return sorted(user_groups)
+
+
+def get_users_in_plone_groups_cachekey(method, group_id=None, group=None, the_objects=False):
+    """cachekey method for self.get_users_in_plone_groups."""
+    date = get_cachekey_volatile('_users_groups_value')
+    return (date,
+            group and group.id or group_id,
+            the_objects)
+
+
+@ram.cache(get_users_in_plone_groups_cachekey)
+def get_users_in_plone_groups(group_id=None, group=None, the_objects=False):
+    """Just return group.getGroupMembers but cached."""
+    if not group and group_id:
+        group = api.group.get(group_id)
+    if group is None:
+        return []
+    members = group.getGroupMembers()
+    if not the_objects:
+        members = [m.id for m in members]
+    return sorted(members)